Renames files for future logic
