+++ tests/test_get_graph_info.py
@@ -1,4 +1,4 @@
-from project.graph import get_graph_info
+from project.task1 import get_graph_info
 import pytest
 
 def test_get_graph_info():
